Moved Hash128 into Hashing namespace
